+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 
-* Make `KeyValueStore::TOKEN` constant public
+* Improve support for custom views.
+* Replace `DoubleRideTimeDataType` sample with `PowerHeartRateDataType` as a better sample.
+* Update `CustomSpeedSdkView` and associated data type as a better sample with external library graphics.
+* Add README link for graphical implementation guidelines.